feat: add interface in use case UnicoCallBackUseCase
