+++ android/build.gradle
@@ -44,6 +44,6 @@ dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
     implementation 'androidx.appcompat:appcompat:1.3.1'
     implementation 'com.google.code.gson:gson:2.8.6'
-    implementation 'com.github.acesso-io:acessobio-android:4.0.15'
+    implementation 'com.github.acesso-io:acessobio-android:4.0.17-beta'
 }
