+++ lib/src/unico/adapter/api/unico.check.builder.default.dart
@@ -7,6 +7,7 @@ import 'package:unico_check/src/unico/domain/entities/unico.config.dart';
 import 'package:unico_check/src/unico/domain/entities/unico.theme.dart';
 import 'package:unico_check/src/unico/domain/usecase/open.camera.usecase.default.dart';
 import 'package:unico_check/src/unico/domain/usecase/unico.callback.usecase.dart';
+import 'package:unico_check/src/unico/domain/usecase/unico.callback.usecase.default.dart';
 import 'package:unico_check/src/unico/plugins/channel/channel.unico.default.dart';
 import 'unico.check.camera.opener.dart';
 
@@ -36,7 +37,7 @@ class UnicoCheck extends UnicoCheckBuilder {
         timeoutSession: _timeoutSession,
         unicoConfigIos: _unicoConfigIos,
         unicoConfigAndroid: _unicoConfigAndroid,
-        unicoCallBackUseCase: UnicoCallBackUseCase());
+        unicoCallBackUseCase: UnicoCallBackUseCaseDefault());
   }
 
   @override
